FIX: Enable SkillWebSocketAdapter dependency injection of IAuthenticator
