+++ sdk/csharp/libraries/microsoft.bot.builder.solutions/Skills/SkillWebSocketAdapter.cs
@@ -22,84 +22,15 @@ namespace Microsoft.Bot.Builder.Solutions.Skills
     /// 2. Create RequestHandler to handle follow-up websocket frames.
     /// 3. Start listening on the websocket connection.
     /// </summary>
-    public class SkillWebSocketAdapter : ISkillWebSocketAdapter
+    public class SkillWebSocketAdapter : SkillWebSocketAdapterBase
     {
-        private readonly IBotTelemetryClient _botTelemetryClient;
-        private readonly SkillWebSocketBotAdapter _skillWebSocketBotAdapter;
-        private readonly IAuthenticator _authenticator;
-        private readonly Stopwatch _stopWatch;
-
         public SkillWebSocketAdapter(
             SkillWebSocketBotAdapter skillWebSocketBotAdapter,
             BotSettingsBase botSettingsBase,
             IWhitelistAuthenticationProvider whitelistAuthenticationProvider,
             IBotTelemetryClient botTelemetryClient = null)
+            : base(skillWebSocketBotAdapter, new MsJWTAuthenticator(botSettingsBase, whitelistAuthenticationProvider), botTelemetryClient)
         {
-            _skillWebSocketBotAdapter = skillWebSocketBotAdapter ?? throw new ArgumentNullException(nameof(skillWebSocketBotAdapter));
-            var botSettings = botSettingsBase ?? throw new ArgumentNullException(nameof(botSettingsBase));
-            var whitelistProvider = whitelistAuthenticationProvider ?? throw new ArgumentNullException(nameof(whitelistAuthenticationProvider));
-            _authenticator = new MsJWTAuthenticator(botSettings, whitelistProvider);
-
-            _botTelemetryClient = botTelemetryClient ?? NullBotTelemetryClient.Instance;
-            _stopWatch = new Stopwatch();
-        }
-
-        public async Task ProcessAsync(HttpRequest httpRequest, HttpResponse httpResponse, IBot bot, CancellationToken cancellationToken = default(CancellationToken))
-        {
-            if (httpRequest == null)
-            {
-                throw new ArgumentNullException(nameof(httpRequest));
-            }
-
-            if (httpResponse == null)
-            {
-                throw new ArgumentNullException(nameof(httpResponse));
-            }
-
-            if (bot == null)
-            {
-                throw new ArgumentNullException(nameof(bot));
-            }
-
-            if (!httpRequest.HttpContext.WebSockets.IsWebSocketRequest)
-            {
-                httpResponse.StatusCode = (int)HttpStatusCode.BadRequest;
-                await httpResponse.WriteAsync("Upgrade to WebSocket required.").ConfigureAwait(false);
-                return;
-            }
-
-            var claimsIdentity = await _authenticator.AuthenticateAsync(httpRequest, httpResponse).ConfigureAwait(false);
-
-            await CreateWebSocketConnectionAsync(claimsIdentity, httpRequest.HttpContext, bot).ConfigureAwait(false);
-        }
-
-        private async Task CreateWebSocketConnectionAsync(ClaimsIdentity claimsIdentity, HttpContext httpContext, IBot bot)
-        {
-            var socket = await httpContext.WebSockets.AcceptWebSocketAsync().ConfigureAwait(false);
-            var handler = new SkillWebSocketRequestHandler(claimsIdentity, _botTelemetryClient);
-            var server = new WebSocketServer(socket, handler);
-            server.Disconnected += Server_Disconnected;
-            _skillWebSocketBotAdapter.Server = server;
-            handler.Bot = bot;
-            handler.SkillWebSocketBotAdapter = _skillWebSocketBotAdapter;
-
-            _botTelemetryClient.TrackTrace("Starting listening on websocket", Severity.Information, null);
-            _stopWatch.Start();
-            var startListening = server.StartAsync();
-            Task.WaitAll(startListening);
-        }
-
-        private void Server_Disconnected(object sender, DisconnectedEventArgs e)
-        {
-            if (_stopWatch.IsRunning)
-            {
-                _stopWatch.Stop();
-
-                _botTelemetryClient.TrackEvent("SkillWebSocketOpenCloseLatency", null, new Dictionary<string, double>
-                {
-                    { "Latency", _stopWatch.ElapsedMilliseconds },
-                });
-            }
         }
     }
 }
